+++ pkgs/swarmauri/swarmauri/llms/concrete/OpenAIToolModel.py
@@ -20,22 +20,22 @@ class OpenAIToolModel(LLMBase):
 
     api_key: str
     allowed_models: List[str] = [
-        "gpt-4o",
-        "gpt-4o-2024-05-13",
-        "gpt-4-turbo",
-        "gpt-4o-mini",
-        "gpt-4o-mini-2024-07-18",
-        "gpt-4o-2024-08-06",
-        "gpt-4-turbo-2024-04-09",
-        "gpt-4-turbo-preview",
-        "gpt-4-0125-preview",
-        "gpt-4-1106-preview",
-        "gpt-4",
-        "gpt-4-0613",
-        "gpt-3.5-turbo",
+        # "gpt-4o",
+        # "gpt-4o-2024-05-13",
+        # "gpt-4-turbo",
+        # "gpt-4o-mini",
+        # "gpt-4o-mini-2024-07-18",
+        # "gpt-4o-2024-08-06",
+        # "gpt-4-turbo-2024-04-09",
+        # "gpt-4-turbo-preview",
+        # "gpt-4-0125-preview",
+        # "gpt-4-1106-preview",
+        # "gpt-4",
+        # "gpt-4-0613",
+        # "gpt-3.5-turbo",
         "gpt-3.5-turbo-0125",
-        "gpt-3.5-turbo-1106",
-        "gpt-3.5-turbo-0613",
+        # "gpt-3.5-turbo-1106",
+        # "gpt-3.5-turbo-0613",
     ]
     name: str = "gpt-3.5-turbo-0125"
     type: Literal["OpenAIToolModel"] = "OpenAIToolModel"
